requirements.txt
amqp==2.5.0
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
decorator==4.4.0
Django==2.2.3
django-cors-headers==2.4.0
django-filter==2.2.0
django-nose==1.4.6
django-prettyjson==0.4.1
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
docker==4.0.2
idna==2.8
ipdb==0.12.1
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
kombu==4.6.3
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
nose==1.3.7
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
requests==2.22.0
sentry-sdk==0.10.2
six==1.12.0
slackclient==1.3.0
sqlparse==0.3.0
standardjson==0.3.1
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
websocket-client==0.56.0
wrapt==1.11.2
